// prisma/generated/type-graphql/models/Content_type.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../scalars";

@TypeGraphQL.ObjectType("Content_type", {
  isAbstract: true
})
export class Content_type {
  @TypeGraphQL.Field(_type => GraphQLScalars.BigIntResolver, {
    nullable: false
  })
  id!: bigint;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  created_at?: Date | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  text?: string | null;

  @TypeGraphQL.Field(_type => DecimalJSScalar, {
    nullable: true
  })
  number?: Prisma.Decimal | null;

  @TypeGraphQL.Field(_type => GraphQLScalars.JSONResolver, {
    nullable: true
  })
  json?: Prisma.JsonValue | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  link?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  email?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  uid?: string | null;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  date?: Date | null;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  time?: Date | null;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  timestamp?: Date | null;

  @TypeGraphQL.Field(_type => Boolean, {
    nullable: true
  })
  boolean?: boolean | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  rich_text?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  password?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  media?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  database_name?: string | null;
}